Consolidated custom variable logic
